% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radb-package.R
\name{radb-exports}
\alias{radb-exports}
\alias{\%>\%}
\title{radb exported operators}
\description{
The following functions are imported and then re-exported
from the radb package to enable use of the magrittr
pipe operator with no additional library calls
}
