% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radb-package.R
\docType{package}
\name{radb}
\alias{radb}
\alias{radb-package}
\title{Tools to Query the 'Merit' 'RADb' Network Route Server}
\description{
'Merit' maintains an active database of metadat and routing information for
the internet. Tools are provided to query this database.
}
\author{
Bob Rudis (bob@rud.is)
}
