% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radb.r
\name{radb_close}
\alias{radb_close}
\title{Close the connection to the Merit RADb server}
\usage{
radb_close(rdb)
}
\arguments{
\item{rdb}{\code{radb} object created with \url{radb_open}}
}
\description{
Close the connection to the Merit RADb server
}
\references{
\url{http://www.radb.net/support/query2.php}
}
