% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radb.r
\name{radb_match}
\alias{radb_match}
\title{Match an object of the specified type with the specified key}
\usage{
radb_match(rdb, key, value)
}
\arguments{
\item{rdb}{\code{radb} object created with \url{radb_open}}

\item{key, value}{what to search by}
}
\description{
Match an object of the specified type with the specified key
}
\references{
\url{http://www.radb.net/support/query2.php}
}
