% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radb.r
\name{radb_open}
\alias{radb_open}
\title{Open a persistent connection to the Merit RADb server}
\usage{
radb_open(timeout = 3600, server = "whois.radb.net")
}
\arguments{
\item{timeout}{timeout in seconds}

\item{server}{server to connect to. Should not need to be modified.}
}
\description{
Open a persistent connection to the Merit RADb server
}
\references{
\url{http://www.radb.net/support/query2.php}
}
