% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radb.r
\name{radb_route_search}
\alias{radb_route_search}
\title{Perform route searches.}
\usage{
radb_route_search(rdb, cidr, option = c("o", "l", "L", "M"))
}
\arguments{
\item{rdb}{\code{radb} object created with \url{radb_open}}

\item{option}{one of "\code{o}", "\code{l}", "\code{L}", "\code{M}". \code{o} - return origin of exact
match(es); \code{l} - one-level less specific; \code{L} - all less specific; \code{M} -
all more specific.}

\item{route}{CIDR"}
}
\description{
Perform route searches.
}
\references{
\url{http://www.radb.net/support/query2.php}
}
