% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radb.r
\name{radb_route_set}
\alias{radb_route_set}
\title{Get as-set or route-set}
\usage{
radb_route_set(rdb, as_set, recursive = FALSE)
}
\arguments{
\item{rdb}{\code{radb} object created with \url{radb_open}}

\item{as_set}{\code{as-set} to retrieve with optional "\code{AS-}" prefix}

\item{recursive}{recursively expand members?}
}
\description{
Get as-set or route-set
}
\references{
\url{http://www.radb.net/support/query2.php}
}
