% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radb.r
\name{radb_routes_by_origin}
\alias{radb_routes_by_origin}
\title{Get routes by origin}
\usage{
radb_routes_by_origin(rdb, asn, ipv6 = FALSE)
}
\arguments{
\item{rdb}{\code{radb} object created with \url{radb_open}}

\item{asn}{autonomous system. Can be a bare integer or prefixed with "\code{AS}" or "\code{as}"}

\item{ipv6}{if \code{TRUE} then get IPv6 routes. Default \code{FALSE}}
}
\description{
Get routes by origin
}
\references{
\url{http://www.radb.net/support/query2.php}
}
