% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radb.r
\name{radb_version}
\alias{radb_version}
\title{Get RADb version}
\usage{
radb_version(rdb)
}
\arguments{
\item{rdb}{\code{radb} object created with \url{radb_open}}
}
\description{
Displays RADb version and adds it to the \code{radb} returned object.
}
\references{
\url{http://www.radb.net/support/query2.php}
}
